requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex==0.1
_openmp_mutex==5.1
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.0.3
antlr-python-runtime==4.9.3
astor==0.8.1
async-timeout==4.0.2
attrs==21.4.0
blas==1.0
blinker==1.4
brotlipy==0.7.0
bzip2==1.0.8
c-ares==1.18.1
ca-certificates==2022.5.18.1
cachetools==5.0.0
cairo==1.16.0
certifi>=2022.12.07
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.0.0
cmake==3.22.1
colorama==0.4.4
contextlib2==21.6.0
cryptography>=38.0.3
cudatoolkit==11.3.1
cycler==0.11.0
cytoolz==0.11.0
dask-core==2022.5.0
dbus==1.13.18
expat==2.4.4
ffmpeg==4.3
filelock==3.7.0
fontconfig==2.13.1
freetype==2.11.0
frozenlist==1.3.0
fsspec==2022.3.0
future>=0.18.3
geos==3.8.0
gettext==0.21.0
giflib==5.2.1
glib==2.68.4
glib-tools==2.68.4
gmp==6.2.1
gnutls==3.6.15
google-auth==2.6.6
google-auth-oauthlib==0.4.6
graphite2==1.3.14
grpcio==1.42.0
gst-plugins-base==1.14.5
gstreamer==1.18.5
harfbuzz==2.7.2
hdf5==1.10.6
hyperopt==0.1.2
icu==67.1
idna==3.3
imageio==2.9.0==pyhd3eb1b0_0
imgaug==0.4.0==pyhd3eb1b0_0
importlib-metadata==4.11.4==py38h578d9bd_0
intel-openmp==2021.4.0==h06a4308_3561
jasper==1.900.1==hd497a04_4
joblib>=1.2.0
jpeg==9e==h7f8727e_0
json-tricks==3.15.5==pypi_0
kiwisolver==1.4.2==py38h295c915_0
krb5==1.19.2==hac12032_0
lame==3.100==h7b6447c_0
lcms2==2.12==h3be6417_0
ld_impl_linux-64==2.38==h1181459_1
libblas==3.9.0==12_linux64_mkl
libcblas==3.9.0==12_linux64_mkl
libclang==11.1.0==default_ha53f305_1
libcurl==7.82.0==h0b77cf5_0
libedit==3.1.20210910==h7f8727e_0
libev==4.33==h7f8727e_1
libevent==2.1.10==h9b69904_4
libffi==3.3==he6710b0_2
libgcc==7.2.0==h69d50b8_2
libgcc-ng==11.2.0==h1234567_0
libgfortran-ng==7.5.0==ha8ba4b0_17
libgfortran4==7.5.0==ha8ba4b0_17
libglib==2.68.4==h3e27bee_0
libgomp==11.2.0==h1234567_0
libiconv==1.16==h7f8727e_2
libidn2==2.3.2==h7f8727e_0
liblapack==3.9.0==12_linux64_mkl
liblapacke==3.9.0==12_linux64_mkl
libllvm11==11.1.0==h3826bc1_1
libnghttp2==1.46.0==hce63b2e_0
libopencv==4.4.0==py38_2
libpng==1.6.37==hbc83047_0
libpq==12.9==hc2c5182_1
libprotobuf==3.15.8==h780b84a_1
libssh2==1.10.0==h8f2d780_0
libstdcxx-ng==11.2.0==h1234567_0
libtasn1==4.16.0==h27cfd23_0
libtiff==4.2.0==h2818925_1
libunistring==0.9.10==h27cfd23_0
libuuid==1.0.3==h7f8727e_2
libuv==1.40.0==h7b6447c_0
libwebp==1.2.2==h55f646e_0
libwebp-base==1.2.2==h7f8727e_0
libxcb==1.15==h7f8727e_0
libxkbcommon==1.0.3==he3ba5ed_0
libxml2==2.9.10==h68273f3_2
locket==1.0.0==py38h06a4308_0
lz4-c==1.9.3==h295c915_1
markdown==3.3.7==pyhd8ed1ab_0
matplotlib==3.2.2==1
matplotlib-base==3.2.2==py38h5d868c9_1
mkl==2021.4.0==h06a4308_640
mkl-service==2.4.0==py38h7f8727e_0
mkl_fft==1.3.1==py38hd3c417c_0
mkl_random==1.2.2==py38h51133e4_0
multidict==6.0.2==py38h0a891b7_1
mysql-common==8.0.25==ha770c72_2
mysql-libs==8.0.25==hfa10184_2
ncurses==6.3==h7f8727e_2
nettle==3.7.3==hbbd107a_1
networkx==2.7.1==pyhd3eb1b0_0
nni==2.7==pypi_0
nspr==4.33==h295c915_0
nss==3.74==h0370c37_0
numpy==1.22.3==py38he7a7128_0
numpy-base==1.22.3==py38hf524024_0
oauthlib>3.2.1==pyhd8ed1ab_0
omegaconf==2.1.2==py38h578d9bd_0
opencv==4.4.0==py38_2
openh264==2.1.1==h4ff587b_0
openssl==1.1.1o==h166bdaf_0
packaging==21.3==pyhd3eb1b0_0
pandas==1.4.2==pypi_0
partd==1.2.0==pyhd3eb1b0_1
pcre==8.45==h295c915_0
pillow>=9.0.1
pip==21.2.4==py38h06a4308_0
pixman==0.40.0==h7f8727e_1
prettytable==3.3.0==pypi_0
protobuf>=3.18.3
psutil==5.9.1==pypi_0
py-opencv==4.4.0==py38h23f93f0_2
pyasn1==0.4.8==py_0
pyasn1-modules==0.2.7==py_0
pycparser==2.21==pyhd3eb1b0_0
pyjwt>=2.4.0
pymongo==4.1.1==pypi_0
pyopenssl==22.0.0==pyhd3eb1b0_0
pyparsing==3.0.4==pyhd3eb1b0_0
pysocks==1.7.1==py38h06a4308_0
python==3.8.13==h12debd9_0
python-dateutil==2.8.2==pyhd3eb1b0_0
python_abi==3.8==2_cp38
pythonwebhdfs==0.2.3==pypi_0
pytorch==1.11.0==py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex==1.0==cuda
pytz==2022.1==pypi_0
pyu2f==0.1.5==pyhd8ed1ab_0
pywavelets==1.3.0==py38h7f8727e_0
pyyaml==6.0==py38h7f8727e_1
qt==5.12.9==h763d07f_1
readline==8.1.2==h7f8727e_1
requests==2.27.1==pyhd3eb1b0_0
requests-oauthlib==1.3.1==pyhd8ed1ab_0
responses==0.20.0==pypi_0
rhash==1.4.1==h3c74f83_1
rsa==4.8==pyhd8ed1ab_0
schema==0.7.5==pypi_0
scikit-image==0.19.2==py38h51133e4_0
scikit-learn==1.1.1==pypi_0
scipy==1.7.3==py38hc147768_0
setuptools>=65.5.1
shapely==1.7.1==py38h1728cc4_0
simplejson==3.17.6==pypi_0
six==1.16.0==pyhd3eb1b0_1
sqlite==3.38.3==hc218d9a_0
tensorboard==2.9.0==pyhd8ed1ab_0
tensorboard-data-server==0.6.0==py38h2b5fc30_2
tensorboard-plugin-wit==1.8.1==pyhd8ed1ab_0
threadpoolctl==3.1.0==pypi_0
tifffile==2020.10.1==py38hdd07704_2
tk==8.6.11==h1ccaba5_1
toolz==0.11.2==pyhd3eb1b0_0
torch-tb-profiler==0.4.0==pypi_0
torchaudio==0.11.0==py38_cu113
torchvision==0.12.0==py38_cu113
tornado==6.1==py38h27cfd23_0
tqdm==4.64.0==pyhd8ed1ab_0
typeguard==2.13.3==pypi_0
typing-extensions==4.1.1==hd3eb1b0_0
typing_extensions==4.1.1==pyh06a4308_0
urllib3==1.26.9==py38h06a4308_0
wcwidth==0.2.5==pypi_0
websockets==10.3==pypi_0
Werkzeug>=2.2.3
wheel>=0.38.1
xz==5.2.5==h7f8727e_1
yaml==0.2.5==h7b6447c_0
yarl==1.7.2==py38h0a891b7_2
zipp==3.8.0==pyhd8ed1ab_0
zlib==1.2.12==h7f8727e_2
zstd==1.5.2==ha4553b6_0
